move sidebar to core
